ADD(lokijs) queryMatcher and sortComparator
