+++ src/plugins/lokijs/index.ts
@@ -1,0 +1,1 @@
+export * from './rx-storage-lokijs';
