+++ test/unit.test.ts
@@ -3,6 +3,7 @@ import './unit/util.test.js';
 import './unit/pouch-db-integration.test.js';
 import './unit/adapter-check.test.js';
 import './unit/rx-storage-pouchdb.test.js';
+import './unit/rx-storage-implementations.test.js';
 import './unit/instance-of-check.test.js';
 import './unit/rx-schema.test.js';
 import './unit/key-compression.test.js';
